chore(lambda layer): move generated props to a dedicated folder
