+++ .projen/files.json
@@ -23,9 +23,9 @@
     ".projen/files.json",
     ".projen/tasks.json",
     "LICENSE",
+    "src/common/props/AdapterProps.ts",
     "src/common/props/DockerLambdaCustomProps.ts",
-    "src/patterns/gen-ai/aws-langchain-common-layer/AdapterProps.ts",
-    "src/patterns/gen-ai/aws-langchain-common-layer/LangchainProps.ts",
+    "src/common/props/LangchainProps.ts",
     "tsconfig.dev.json"
   ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.ts and run \"npx projen\"."
