+++ .projenrc.ts
@@ -169,12 +169,12 @@ project.eslint?.addIgnorePattern('LangchainProps.ts');
 project.eslint?.addIgnorePattern('AdapterProps.ts');
 
 // Shared interfaces extending pre-existing CDK interfaces
-new ProjenStruct(project, { name: 'LangchainProps', filePath: 'src/patterns/gen-ai/aws-langchain-common-layer/LangchainProps.ts' })
+new ProjenStruct(project, { name: 'LangchainProps', filePath: 'src/common/props/LangchainProps.ts' })
   .mixin(Struct.fromFqn('aws-cdk-lib.aws_lambda.LayerVersionProps'))
   .withoutDeprecated()
   .omit('code', 'compatibleRuntimes', 'compatibleArchitectures');
 
-new ProjenStruct(project, { name: 'AdapterProps', filePath: 'src/patterns/gen-ai/aws-langchain-common-layer/AdapterProps.ts' })
+new ProjenStruct(project, { name: 'AdapterProps', filePath: 'src/common/props/AdapterProps.ts' })
   .mixin(Struct.fromFqn('aws-cdk-lib.aws_lambda.LayerVersionProps'))
   .withoutDeprecated()
   .omit('code');
