+++ src/common/props/AdapterProps.ts
